FEATURE: Better support for custom select elements
Adjusts the `Neos.Form.Builder:NodeBasedFormElement` Fusion prototype
slightly to use processors for `properties.options` and `elements` in
order to make it easier to manipulate those in custom implementations.

This also adds documentation on how to create custom select elements.

Background:

When using `@if` alone on a property, its value will be `null` if the
condition doesn't match - even if the property has a value in the
extended prototype.