BUGFIX: Allow `NodeBasedFormElement`s to be overridden
The `properties.options` of a form element were overridden
by the `NodeBasedFormElement` prototype if the node implemented
the `Neos.Form.Builder:SelectionMixin`.
Otherwise the `@if` condition led to an *empty* array even
if it was filled in the concrete implementation.

By replacing the implementation with a `@process` expression
the value is untouched if the condition is not satisfied, allowing
custom implementations to override the value.

This also fixes the same issue for the `elements` option of
`SectionMixin` nodes.

Fixes: #18