Enable support for custom Form Builder Styling with the new Neos UI
This adjusts the content element wrapping of all node based forms
and form elements so they have a custom attribute `data-_neos-form-builder-type`
with a value of the corresponding node type.
In addition empty content collections get a `data-_neos-form-builder-empty-collection`
attribute.

This allows for custom styling of the builder in the ReactJS based
Neos backend as described in the README.

Resolves: #24